Annotation update

+++ annotations/nocedit.exe/pseudocode/include/types/classes/CZThumb.h
@@ -11,7 +11,7 @@ typedef struct CZThumb {
     int height; // 0x4
     CVector3f camera_position; // 0x8
     CVector3f look_at_position; // 0x14
-    int projection_scale; // 0x20
+    float projection_scale; // 0x20
     void* zbuffer_data; // 0x24
 } CZThumb;
